+++ src/endstone/core/block/block.h
@@ -23,9 +23,9 @@ namespace endstone::core {
 class EndstoneBlock : public Block {
 public:
     EndstoneBlock(BlockSource &block_source, BlockPos block_pos);
-    [[nodiscard]] std::string getType() const override;
-    void setType(std::string type) override;
-    void setType(std::string type, bool apply_physics) override;
+    [[nodiscard]] const BlockType &getType() const override;
+    void setType(BlockTypeId type) override;
+    void setType(BlockTypeId type, bool apply_physics) override;
     [[nodiscard]] std::unique_ptr<BlockData> getData() const override;
     void setData(const BlockData &data) override;
     void setData(const BlockData &data, bool apply_physics) override;