+++ src/endstone/core/block/block_data.h
@@ -23,7 +23,7 @@ class EndstoneBlockData : public BlockData {
 public:
     explicit EndstoneBlockData(::Block &block);
 
-    [[nodiscard]] std::string getType() const override;
+    [[nodiscard]] const BlockType &getType() const override;
     [[nodiscard]] std::string getTranslationKey() const override;
     [[nodiscard]] BlockStates getBlockStates() const override;
     [[nodiscard]] std::uint32_t getRuntimeId() const override;